+++ CondTools/RPC/src/RPCTempSH.cc
@@ -1,8 +1,8 @@
 /*
  *  See headers for a description
  *
- *  $Date: 2008/12/30 10:11:13 $
- *  $Revision: 1.4 $
+ *  $Date: 2009/06/05 14:33:28 $
+ *  $Revision: 1.6 $
  *  \author D. Pagano - Dip. Fis. Nucl. e Teo. & INFN Pavia
  */
 
@@ -18,8 +18,8 @@ popcon::RpcDataT::RpcDataT(const edm::ParameterSet& pset) :
   host(pset.getUntrackedParameter<std::string>("host", "source db host")),
   user(pset.getUntrackedParameter<std::string>("user", "source username")),
   passw(pset.getUntrackedParameter<std::string>("passw", "source password")),
-  m_since(pset.getUntrackedParameter<unsigned long long>("since",5)),
-  m_till(pset.getUntrackedParameter<unsigned long long>("till",0)){
+  m_since(pset.getUntrackedParameter<uint64_t>("since",5)),
+  m_till(pset.getUntrackedParameter<uint64_t>("till",0)){
 }
 
 popcon::RpcDataT::~RpcDataT()
